리팩토링 - InputContainer에 Input 관련 상태관리 추출
- Input에 대한 상태 관리를 App에서 InputContainer로 리팩토링
- InputContainer 테스트 추가
