+++ src/Page.test.jsx
@@ -1,4 +1,4 @@
-import { render, fireEvent } from '@testing-library/react';
+import { render } from '@testing-library/react';
 
 import { useSelector } from 'react-redux';
 
@@ -14,21 +14,10 @@ test('Page', () => {
     ],
   }));
 
-  const handleChangeTitle = jest.fn();
-  const handleClickAddTask = jest.fn();
-
   const { getByText } = render((
-    <Page
-      taskTitle=""
-      onChangeTitle={handleChangeTitle}
-      onClickAddTask={handleClickAddTask}
-    />
+    <Page />
   ));
 
   expect(getByText(/Task-1/)).not.toBeNull();
   expect(getByText(/Task-2/)).not.toBeNull();
-
-  fireEvent.click(getByText('추가'));
-
-  expect(handleClickAddTask).toBeCalled();
 });
